+++ build.gradle.kts
@@ -42,7 +42,7 @@ dependencies {
     dokka(projects.exposed.exposedSpringBoot4Starter)
     dokka(projects.exposed.springTransaction)
     dokka(projects.exposed.spring7Transaction)
-    dokka(projects.exposed.springReactiveTransaction)
+    dokka(projects.exposed.spring7ReactiveTransaction)
 
     // Kover aggregated coverage dependencies
     // Include all source modules for coverage aggregation
@@ -55,7 +55,7 @@ dependencies {
     kover(project(":exposed-java-time"))
     kover(project(":spring-transaction"))
     kover(project(":spring7-transaction"))
-    kover(project(":spring-reactive-transaction"))
+    kover(project(":spring7-reactive-transaction"))
     kover(project(":exposed-spring-boot-starter"))
     kover(project(":exposed-spring-boot4-starter"))
     kover(project(":exposed-jdbc"))
