+++ settings.gradle.kts
@@ -7,7 +7,7 @@ include("exposed-jodatime")
 include("exposed-java-time")
 include("spring-transaction")
 include("spring7-transaction")
-include("spring-reactive-transaction")
+include("spring7-reactive-transaction")
 include("exposed-spring-boot-starter")
 include("exposed-spring-boot4-starter")
 include("exposed-jdbc")
