--- spring-reactive-transaction/build.gradle.kts
@@ -1,59 +1,0 @@
-import org.gradle.api.tasks.testing.logging.*
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.tasks.*
-
-plugins {
-    kotlin("jvm")
-
-    alias(libs.plugins.dokka)
-}
-
-repositories {
-    mavenCentral()
-}
-
-kotlin {
-    jvmToolchain(17)
-}
-
-dependencies {
-    api(project(":exposed-core"))
-    api(project(":exposed-r2dbc"))
-    api(libs.spring.r2dbc)
-    api(libs.spring7.context)
-    implementation(libs.kotlinx.coroutines.reactor)
-
-    testImplementation(project(":exposed-r2dbc-tests"))
-    testImplementation(kotlin("test"))
-    testImplementation(libs.junit6)
-    testRuntimeOnly(libs.junit.platform.launcher)
-    testImplementation(libs.kotlinx.coroutines.debug)
-    testImplementation(libs.kotlinx.coroutines.test)
-    testImplementation(libs.spring7.test)
-    testImplementation(libs.slf4j)
-    testImplementation(libs.log4j.slf4j.impl)
-    testImplementation(libs.log4j.api)
-    testImplementation(libs.log4j.core)
-    testImplementation(libs.r2dbc.h2) {
-        exclude(group = "com.h2database", module = "h2")
-    }
-}
-
-tasks.withType<KotlinCompile>().configureEach {
-    compilerOptions {
-        jvmTarget.set(JvmTarget.JVM_17)
-    }
-}
-
-tasks.withType<Test>().configureEach {
-    if (JavaVersion.VERSION_1_8 > JavaVersion.current()) {
-        jvmArgs = listOf("-XX:MaxPermSize=256m")
-    }
-    testLogging {
-        events.addAll(listOf(TestLogEvent.PASSED, TestLogEvent.FAILED, TestLogEvent.SKIPPED))
-        showStandardStreams = true
-        exceptionFormat = TestExceptionFormat.FULL
-    }
-
-    useJUnitPlatform()
-}
